Hack together the beginnings of the compiler
I wanted to be able to print to the console by the end of the stream, so
this is a poorly fixed-up hack to make that work. The next few days will
be dedicated to refactoring and expanding the existing systems to work
nicely for Score in the future.

I can't wait to write the compiler in Score...